# 02/app/api/swagger.yaml
openapi: 3.0.0
info:
  title: User Service
  description: |
    This is simple client API
  contact:
    email: schetinnikov@gmail.com
  version: 1.0.0
servers:
- url: http://example.domain/api/v1/
  description: Example Server
- url: https://virtserver.swaggerhub.com/otus55/users/1.0.0
  description: SwaggerHub API Auto Mocking
tags:
- name: user
  description: Operations about user
paths:
  /user:
    post:
      tags:
      - user
      summary: Create user
      description: This can only be done by the logged in user.
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
            examples:
              sample-user:
                summary: Example
                value:
                  username: johndoe589
                  firstName: John
                  lastName: Doe
                  email: bestjohn@doe.com
                  phone: +71002003040
        required: true
      responses:
        default:
          description: successful operation
  /user/{userId}:
    get:
      tags:
      - user
      description: Returns a user based on a single ID, if the user does not have
        access to the user
      operationId: find user by id
      parameters:
      - name: userId
        in: path
        description: ID of user
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: user response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
      - user
      description: Update user with User ID supplied
      operationId: updateUser
      parameters:
      - name: userId
        in: path
        description: ID of user
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
            examples:
              sample-user:
                summary: Example
                value:
                  firstName: Julie
                  lastName: Doe
                  email: bestjohn@doe.com
                  phone: +71004242424
      responses:
        "200":
          description: user updated
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
      - user
      description: deletes a single user based on the ID supplied
      operationId: deleteUser
      parameters:
      - name: userId
        in: path
        description: ID of user
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: user deleted
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          maxLength: 256
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          format: phone
      example:
        firstName: firstName
        lastName: lastName
        phone: phone
        id: 0
        email: ""
        username: username
    Error:
      required:
      - code
      - message
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
  requestBodies:
    UserArray:
      description: List of user object
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/User'
      required: true